pyproject.toml
[project]
name = "mcp-clickhouse"
version = "0.1.4"
description = "An MCP server for ClickHouse."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.13"
dependencies = [
     "mcp[cli]>=1.3.0",
     "python-dotenv>=1.0.1",
     "uvicorn>=0.34.0",
     "clickhouse-connect>=0.8.0",
     "pip-system-certs>=4.0",
]

[project.scripts]
mcp-clickhouse = "mcp_clickhouse.main:main"

[project.urls]
Home = "https://github.com/ClickHouse/mcp-clickhouse"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest"
]

[tool.hatch.build.targets.wheel]
packages = ["mcp_clickhouse"]

[tool.ruff]
line-length = 100

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
